jinja_report/requirements.txt
Jinja2>=2.10
MarkupSafe>=1.0
pyyaml>=6.0
pandas>=1.4.4